More flight tests. looking good.  300hz ESC refresh rate tested and works, maybe a little more sensitive to the PID values, going back to the 250hz.

+++ FlightController/FlightController.ino
@@ -25,7 +25,7 @@ int yaw_setpoint      = 0;
 int throttle = MIN_ESC_SIGNAL;
 
 double dt = 0.0;
-double timer = 0.0;
+unsigned long  timer = 0;
 
 #include "MPU.h"
 #include "PID.h"
